+++ internal/nest/nest.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"slices"
 
 	"github.com/mhemeryck/nest/internal/config"
 	"github.com/mhemeryck/nest/internal/controller"
@@ -90,38 +89,4 @@ func Run(ctx context.Context, opts Options) error {
 
 	slog.Info("shutting down")
 	return nil
-}
-
-func mqttConfigFromEntity(cfg entity.MQTT) nestmqtt.Config {
-	return nestmqtt.Config{
-		Enabled:         cfg.Enabled,
-		Broker:          cfg.Broker,
-		UnitID:          cfg.UnitID,
-		ClientID:        cfg.ClientID,
-		Username:        cfg.Username,
-		Password:        cfg.Password,
-		DiscoveryPrefix: cfg.DiscoveryPrefix,
-	}
-}
-
-func configuredDevices(devices []*sysfs.Device, wanted []entity.DeviceID) ([]*sysfs.Device, []entity.DeviceID) {
-	byIdentifier := make(map[string]*sysfs.Device, len(devices))
-	for _, device := range devices {
-		byIdentifier[device.Identifier] = device
-	}
-
-	configured := make([]*sysfs.Device, 0, len(wanted))
-	missing := make([]entity.DeviceID, 0)
-	for _, identifier := range wanted {
-		device, ok := byIdentifier[string(identifier)]
-		if !ok {
-			missing = append(missing, identifier)
-			continue
-		}
-		configured = append(configured, device)
-	}
-
-	slices.Sort(missing)
-
-	return configured, missing
 }
